+++ CHANGELOG.md
@@ -18,10 +18,10 @@
 * **o-language-selector**: new component for changing application locale listing all available languages. It  uses '*flag-icons*' library ([d7b0c25](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/d7b0c25))
 * **o-app-layout**: adding the following inputs:
   * '*sidenav-opened*', '*show-header*', '*show-user-info*', '*opened-sidenav-image*' and '*closed-sidenav-image*'
- ([03d79fe](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/03d79fe)) ([ecc3404](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/ecc3404))
+ ([03d79fe](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/03d79fe)) ([ecc3404](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/ecc3404)) ([2cf1de2](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/2cf1de2))
 
   This changes are propagated to inner components ('*o-app-sidenav*', '*o-app-sidenav-image*', '*o-app-sidenav-menu-group*' and '*o-app-sidenav-menu-item*')
-
+* **o-table**: adding input *fixed-header*  ([a0601a6](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/a0601a6)) 
 ### Bug Fixes
 * **o-translate-service**: fixing *'getBrowserLang'* bug ([#138](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/138)) ([4a5d2e7](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/4a5d2e7))
 * **o-form**, **o-table**, **o-datatable**, **o-list**,**o-list-picker**,**o-combo**: fixing to show server messages in the client when the requested action fails in the components bug ([#59](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/59)) ([bceae37](https://github.com/OntimizeWeb/ontimize-web-ngx-datatable/commit/bceae37)) ([6ed6bda](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/6ed6bda))